int to long

+++ bvc/lhsRhsPass.cpp
@@ -2,23 +2,23 @@ using namespace llvm;
 class LhsRhsFinder
 {
     protected:
-        using lhs = std::pair<Value*, int>;
-	    using rhs = std::vector<std::pair<Value*, int>>;
+        using lhs = std::pair<Value*, long long int >;
+	    using rhs = std::vector<std::pair<Value*, long long int >>;
         std::map<Instruction*, std::pair<lhs, rhs>> lhsRhsMap;
-        std::stack<std::vector<std::pair<Value*, int>>> st;
+        std::stack<std::vector<std::pair<Value*, long long int >>> st;
     	std::map<Instruction*, rhs> useMap;
-    	std::map<std::tuple<Function*, BasicBlock*, Instruction*>, std::set<std::tuple<Function*, BasicBlock*, Instruction*>>> predIns, succIns;
-    	std::map<std::pair<BasicBlock*, Function*>, std::map<Instruction*, int>> instCounter;
+    	 std::map<std::tuple<Function*, BasicBlock*, Instruction*>, std::set<std::tuple<Function*, BasicBlock*, Instruction*>>> predIns, succIns;
+    	std::map<std::pair<BasicBlock*, Function*>, std::map<Instruction*, long long int >> instCounter;
     	std::map<Instruction*, rhs> funcToArgMap;
 
     public:
         void metaDataSetter(Function* F);
         void setPredSucc(Function* F);
         void computePredSucc(Instruction* I, BasicBlock* BB, Function* F);
-        std::pair<Value*, int> getLHS(Instruction* ins);
-        std::vector<std::pair<Value*, int>> getRHS(Instruction*);
-        std::vector<std::pair<Value*, int>> getUSE(Instruction*);
-        std::vector<std::pair<Value*, int>> getArg(Instruction*);
+        std::pair<Value*, long long int > getLHS(Instruction* ins);
+        std::vector<std::pair<Value*, long long int >> getRHS(Instruction*);
+        std::vector<std::pair<Value*, long long int >> getUSE(Instruction*);
+        std::vector<std::pair<Value*, long long int >> getArg(Instruction*);
         std::set<std::tuple<Function*, BasicBlock*, Instruction*>> getPredecessor(Function*, BasicBlock*, Instruction*);
         std::set<std::tuple<Function*, BasicBlock*, Instruction*>> getSuccesor(Function*, BasicBlock*, Instruction*);
         void print();
@@ -31,7 +31,7 @@ class LhsRhsFinder
 //------------------ utility function to compute pred and succ ----------//
 void LhsRhsFinder::computePredSucc(Instruction* I, BasicBlock* BB, Function* F)
 {
-    int counter = instCounter[std::make_pair(BB, F)][I];
+    long long int  counter = instCounter[std::make_pair(BB, F)][I];
     if(counter == 0)
     {//first instruction of basic block
         for(auto it = pred_begin(BB), et = pred_end(BB); it != et; ++it)
@@ -67,22 +67,22 @@ std::set<std::tuple<Function*, BasicBlock*, Instruction*>> LhsRhsFinder::getSucc
 
 
 //-------------------- different API ------------------------//
-std::pair<Value*, int> LhsRhsFinder::getLHS(Instruction* ins)
+std::pair<Value*, long long int > LhsRhsFinder::getLHS(Instruction* ins)
 {
     return lhsRhsMap[ins].first;
 }
 
-std::vector<std::pair<Value*, int>> LhsRhsFinder::getRHS(Instruction* ins)
+std::vector<std::pair<Value*, long long int >> LhsRhsFinder::getRHS(Instruction* ins)
 {
     return lhsRhsMap[ins].second;
 }
 
-std::vector<std::pair<Value*, int>> LhsRhsFinder::getUSE(Instruction* ins)
+std::vector<std::pair<Value*, long long int >> LhsRhsFinder::getUSE(Instruction* ins)
 {
     return useMap[ins];
 }
 
-std::vector<std::pair<Value*, int>> LhsRhsFinder::getArg(Instruction* ins)
+std::vector<std::pair<Value*, long long int >> LhsRhsFinder::getArg(Instruction* ins)
 {
     return funcToArgMap[ins];
 }
@@ -98,7 +98,7 @@ void LhsRhsFinder::metaDataSetter(Function* F)
     for(Function::iterator bb=F->begin(), e=F->end(); e!=bb;++bb)
     {
         BasicBlock* BB = &(*bb);
-        int counter = 0;
+        long long int  counter = 0;
         for(BasicBlock::iterator i=bb->begin(), e=bb->end(); i!=e; ++i)
         {
             Instruction* I = &(*i);
@@ -106,8 +106,8 @@ void LhsRhsFinder::metaDataSetter(Function* F)
             {
             //lhs = rhs
                 bool flag = true;
-                std::pair<Value*, int> LHS;
-                std::vector<std::pair<Value*, int>> RHS;
+                std::pair<Value*, long long int > LHS;
+                std::vector<std::pair<Value*, long long int >> RHS;
                 if(I->getOperand(0)->getName().size()>0 and isa<GlobalVariable>(I->getOperand(0)))
                 {
                     RHS.push_back(std::make_pair(I->getOperand(0), -1));
@@ -154,7 +154,7 @@ void LhsRhsFinder::metaDataSetter(Function* F)
             {
                 if(st.empty() or isa<GlobalVariable>(I->getOperand(0)))
                 {
-                    std::vector<std::pair<Value*, int>> v;
+                    std::vector<std::pair<Value*, long long int >> v;
                     v.push_back(std::make_pair(i->getOperand(0), 0));
                     st.push(v);
                 }
@@ -193,7 +193,7 @@ void LhsRhsFinder::metaDataSetter(Function* F)
             else if(st.size() >= 2)
             {
 
-                std::vector<std::pair<Value*, int>> v = st.top();
+                std::vector<std::pair<Value*, long long int >> v = st.top();
                 st.pop();
                 for(auto it : st.top())
                 {
@@ -223,7 +223,7 @@ void LhsRhsFinder::setPredSucc(Function* F)
 
 
 
-//-------------------- print function --------------------------//
+//-------------------- prlong long int  function --------------------------//
 void LhsRhsFinder::print()
 {
     for(auto func: funcToArgMap)
